backend/requirements.txt
fastapi
uvicorn
pydantic[dotenv]
pydantic-settings
requests
pyjwt
passlib[bcrypt]
python-multipart
python-dotenv
